requirements.txt
streamlit
pandas
numpy
requests
python-dotenv
openpyxl

tensorflow
scikit-learn
xgboost
